added get_next_line & ft_printf with all modified functions. Changed also all size_t to int, i may revert it later

+++ ft_calloc.c
@@ -3,17 +3,17 @@
 /*                                                        :::      ::::::::   */
 /*   ft_calloc.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: rpinheir <rpinheir@student.42.fr>          +#+  +:+       +#+        */
+/*   By: rpinheir <rpinheir@student.42lausanne.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/03 12:31:03 by rpinheir          #+#    #+#             */
-/*   Updated: 2025/10/03 13:08:55 by rpinheir         ###   ####lausanne.ch   */
+/*   Updated: 2025/11/26 14:19:51 by rpinheir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 #include <stdlib.h>
 
-void	*ft_calloc(size_t nmemb, size_t size)
+void	*ft_calloc(int nmemb, int size)
 {
 	void	*ptr;
